Introduce pacman.py for pacman related logic
